Extract glTF material extensions from glb parser into separate files
